Add files via upload
A detailed clinical feature dataset of various patients affected by breast cancer. These features include, but are not limited to, age, BMI, insulin, and adiponectin levels. The target variable of this experiment was generating an accurate machine learning model that could predict any given patient's tumor, with any configuration of clinical data. This is synonymous to to deriving a probability density function with a multivariate domain.